refactoring libs appropriately
