+++ src/deeplearning/utils/pipelines.py
@@ -1,11 +1,9 @@
-from contextlib2 import contextmanager
 from ml_metadata.proto import metadata_store_pb2  # type: ignore
 from typing import Optional
 
 import logging
 import numpy  # type: ignore
 import os
-import tempfile as tmp
 import tensorflow as tf  # type: ignore
 import tfx.v1 as tfx  # type: ignore
 
@@ -166,41 +164,6 @@ def _int64_feature(value):
         A tf.train.Feature object containing a tf.train.Int64List.
     '''
     return tf.train.Feature(int64_list=tf.train.Int64List(value=[value]))
-
-
-@contextmanager
-def get_tmp_dir(prefix: str, clean: bool = True):
-    '''
-    Creates a temporary directory and then cleans it up.
-
-    Args:
-        prefix: A string specifying the prefix to use for the temp directory name.
-        clean: A flag to specify whether to clean up the directory on exit.
-
-    Returns:
-        directory: A string specifying the created directory.
-
-    Raises:
-        e (Exception): Any unhandled exception, as necessary.
-    '''
-
-    directory = tmp.mkdtemp(prefix=prefix)
-    try:
-        logger.info(f'Created temporary directory {directory}.')
-        yield directory
-    except Exception as e:
-        logger.exception(e)
-        raise e
-    finally:
-        if clean:
-            for file in os.listdir(directory):
-                path = os.path.join(directory, file)
-                if os.path.isfile(path):
-                    os.remove(path)
-            if len(os.listdir(directory)) == 0:
-                os.rmdir(directory)
-            else:
-                logger.warning(f'Cloud not clean up non-empty directory {directory}.')
 
 
 def serialize_array(array):
